requirements.txt
fastapi==0.121.1
starlette==0.49.3
uvicorn[standard]==0.38.0
watchfiles==1.1.1
sqlalchemy==2.0.44
asyncpg==0.30.0
pydantic==2.12.4
pydantic-settings==2.12.0
python-multipart==0.0.20
python-jose==3.5.0
passlib==1.7.4
bcrypt==5.0.0
aiofiles==25.1.0
httpx==0.28.1
redis==7.0.1
celery==5.5.3
structlog==25.5.0
slowapi==0.1.9
opencv-python==4.12.0.88
pillow==10.4.0
torch==2.3.1
torchvision==0.18.1
git+https://github.com/facebookresearch/segment-anything.git
supervision==0.26.1
uvloop==0.22.1
websockets==15.0.1
greenlet==3.2.4
tenacity==9.1.2
aiohttp==3.13.2
